Fix new clippy lints for Rust 1.66
- https://rust-lang.github.io/rust-clippy/master/index.html#unnecessary_cast
- https://rust-lang.github.io/rust-clippy/master/index.html#manual_clamp
- https://rust-lang.github.io/rust-clippy/master/#suboptimal_flops

+++ src/asm/x86/predict.rs
@@ -220,8 +220,8 @@ pub fn dispatch_predict_intra<T: Pixel>(
               );
               // From dav1d, dx and dy are the distance from the predicted block to the frame edge
               let (dx, dy) = (
-                (bw as isize - dst.rect().x as isize) as libc::c_int,
-                (bh as isize - dst.rect().y as isize) as libc::c_int,
+                (bw as isize - dst.rect().x) as libc::c_int,
+                (bh as isize - dst.rect().y) as libc::c_int,
               );
 
               if angle <= 90 {
@@ -371,8 +371,8 @@ pub fn dispatch_predict_intra<T: Pixel>(
             );
             // From dav1d, dx and dy are the distance from the predicted block to the frame edge
             let (dx, dy) = (
-              (bw as isize - dst.rect().x as isize) as libc::c_int,
-              (bh as isize - dst.rect().y as isize) as libc::c_int,
+              (bw as isize - dst.rect().x) as libc::c_int,
+              (bh as isize - dst.rect().y) as libc::c_int,
             );
 
             if angle <= 90 {
